add personal and friends' donation history tabs
