+++ app/src/main/res/layout/activity_donation_history.xml
@@ -1,26 +1,30 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".NavigationDrawerActivities.DonationHistoryActivity">
+xmlns:app="http://schemas.android.com/apk/res-auto"
+xmlns:tools="http://schemas.android.com/tools"
+android:layout_width="match_parent"
+android:layout_height="match_parent"
+tools:context=".NavigationDrawerActivities.DonationHistoryActivity">
 
     <include
         android:id="@+id/donationHistoryItemToolbar"
-        layout="@layout/toolbar"></include>
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rvDonations"
-        android:layout_width="@dimen/no_layout_width_textview"
-        android:layout_height="@dimen/no_layout_height_textview"
-        android:layout_marginStart="@dimen/layout_margin_start"
-        android:layout_marginLeft="@dimen/layout_margin_start"
-        android:layout_marginEnd="@dimen/layout_margin_end"
-        android:layout_marginRight="@dimen/layout_margin_right"
-        android:layout_marginBottom="@dimen/layout_margin_bottom"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
+        layout="@layout/toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content" />
+
+    <androidx.viewpager.widget.ViewPager
+        android:id="@+id/vpPager"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
         android:layout_marginTop="?attr/actionBarSize"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="1.0"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/donationHistoryItemToolbar">
+
+        <com.google.android.material.tabs.TabLayout
+            android:id="@+id/tab_layout"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content" />
+    </androidx.viewpager.widget.ViewPager>
 </androidx.constraintlayout.widget.ConstraintLayout>
